Needs to reorder the code
