--- CharDevRustV4/.Module.symvers.cmd
@@ -1,1 +1,0 @@
-savedcmd_/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/Module.symvers :=  scripts/mod/modpost -M  -a      -o /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/Module.symvers -T /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/modules.order -i Module.symvers -e
