--- CharDevRustV4/.sec_module.ko.cmd
@@ -1,1 +1,0 @@
-savedcmd_/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/sec_module.ko := ld.lld -r -m elf_x86_64 -z noexecstack --build-id=sha1  -T scripts/module.lds -o /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/sec_module.ko /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/sec_module.o /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/sec_module.mod.o
