+++ CharDevRustV4/Kbuild
@@ -3,4 +3,4 @@
 obj-m := sec_module.o 
 
 # Add dependencies for rust_kprobes
-sec_module-objs := src/sec_module.o c/sec_device.o
+sec_module-objs := src/sec_module.o  c/sec_device.o
