--- CharDevRustV4/sec_module.mod.c
@@ -1,38 +1,0 @@
-#include <linux/module.h>
-#define INCLUDE_VERMAGIC
-#include <linux/build-salt.h>
-#include <linux/elfnote-lto.h>
-#include <linux/export-internal.h>
-#include <linux/vermagic.h>
-#include <linux/compiler.h>
-
-#ifdef CONFIG_UNWINDER_ORC
-#include <asm/orc_header.h>
-ORC_HEADER;
-#endif
-
-BUILD_SALT;
-BUILD_LTO_INFO;
-
-MODULE_INFO(vermagic, VERMAGIC_STRING);
-MODULE_INFO(name, KBUILD_MODNAME);
-
-__visible struct module __this_module
-__section(".gnu.linkonce.this_module") = {
-	.name = KBUILD_MODNAME,
-	.init = init_module,
-#ifdef CONFIG_MODULE_UNLOAD
-	.exit = cleanup_module,
-#endif
-	.arch = MODULE_ARCH_INIT,
-};
-
-#ifdef CONFIG_MITIGATION_RETPOLINE
-MODULE_INFO(retpoline, "Y");
-#endif
-
-
-MODULE_INFO(depends, "");
-
-
-MODULE_INFO(srcversion, "1A3BB33DDB055FEA0CBBF88");
