+++ CharDevRustV4/src/.sec_module.o.d
@@ -1,0 +1,16 @@
+/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/sec_module.o: /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/sec_module.rs /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/structures/mod.rs /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/structures/constant.rs /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/ioctl.rs /home/rustxave/src/kernel/linux/rust/libcore.rmeta /home/rustxave/src/kernel/linux/rust/libcompiler_builtins.rmeta /home/rustxave/src/kernel/linux/rust/libkernel.rmeta /home/rustxave/src/kernel/linux/rust/liballoc.rmeta /home/rustxave/src/kernel/linux/rust/libmacros.so /home/rustxave/src/kernel/linux/rust/libbindings.rmeta /home/rustxave/src/kernel/linux/rust/libuapi.rmeta /home/rustxave/src/kernel/linux/rust/libbuild_error.rmeta
+
+/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/.sec_module.o.d: /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/sec_module.rs /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/structures/mod.rs /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/structures/constant.rs /home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/ioctl.rs /home/rustxave/src/kernel/linux/rust/libcore.rmeta /home/rustxave/src/kernel/linux/rust/libcompiler_builtins.rmeta /home/rustxave/src/kernel/linux/rust/libkernel.rmeta /home/rustxave/src/kernel/linux/rust/liballoc.rmeta /home/rustxave/src/kernel/linux/rust/libmacros.so /home/rustxave/src/kernel/linux/rust/libbindings.rmeta /home/rustxave/src/kernel/linux/rust/libuapi.rmeta /home/rustxave/src/kernel/linux/rust/libbuild_error.rmeta
+
+/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/sec_module.rs:
+/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/structures/mod.rs:
+/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/structures/constant.rs:
+/home/rustxave/Scrivania/Rust-Modules/Rust4Linux/CharDevRustV4/src/ioctl.rs:
+/home/rustxave/src/kernel/linux/rust/libcore.rmeta:
+/home/rustxave/src/kernel/linux/rust/libcompiler_builtins.rmeta:
+/home/rustxave/src/kernel/linux/rust/libkernel.rmeta:
+/home/rustxave/src/kernel/linux/rust/liballoc.rmeta:
+/home/rustxave/src/kernel/linux/rust/libmacros.so:
+/home/rustxave/src/kernel/linux/rust/libbindings.rmeta:
+/home/rustxave/src/kernel/linux/rust/libuapi.rmeta:
+/home/rustxave/src/kernel/linux/rust/libbuild_error.rmeta:
