+++ CharDevRustV4/src/lib.rs
@@ -1,2 +1,4 @@
+// src/lib.rs
 pub mod sec_module;
-pub mod structures;
+
+
