+++ CharDevRustV4/src/structures/constant.rs
@@ -1,0 +1,3 @@
+pub const RULE_SIZE: usize = 256; // Max size of each rule string
+pub const RULE_NUMBER: usize = 16;
+pub const RULE_BUFFER_SIZE: usize = RULE_SIZE * RULE_NUMBER; // Max size of each rule string
